add: refacory CustomWebApplication, refactory tests project, add RegistroAtendimentoDocenteClassFixture

+++ src/RegistroAtendimentoDocente.Domain/Entities/Atendimento.cs
@@ -1,7 +1,7 @@
 ﻿namespace RegistroAtendimentoDocente.Domain.Entities;
 public class Atendimento
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Docente { get; set; } = string.Empty;
     public string Assunto { get; set; } = string.Empty;
     public DateTime Data {  get; set; }
